Update DisplayFloorPlanPanel.dart
Adding in some inline Javascript (and dynamic HTML/svg editing functions) to somewhat enable amenity visibility toggling. However, we might need to edit the 'visibility' attributes for a different level of g tags, because this currently only gets rid of the inner icon, not the outer bubble.

Separately, we also need to style all of this.